mise au propre blt read
